srv/requirements.txt
argon2-cffi==19.1.0
autopep8==1.4.4
backcall==0.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Cython==0.29.13
decorator==4.4.0
Django==2.2.4
django-debug-toolbar==2.0
django-filter==2.2.0
django-guardian==2.0.0
django-nine==0.2.2
django-redis==4.10.0
djangorestframework==3.10.2
entrypoints==0.3
flake8==3.7.8
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.1
lorem==0.1.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.4
names==0.3.0
numpy==1.17.0
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.9
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pytz==2019.2
redis==3.3.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
